src/requirements.txt
# Core framework (Djongo-compatible)
Django==3.1.12
djangorestframework==3.12.4
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.7.2
sqlparse==0.2.4

# Database (required by Djongo)
djongo==1.3.7
pymongo==3.11.4

# Celery / messaging
celery==5.3.6
kombu==5.3.4
amqp==5.2.0
billiard==4.2.0
vine==5.1.0
redis==5.0.1

# Scientific + numeric (ARM64 wheels)
numpy==1.22.4
pandas==1.5.3
matplotlib==3.7.3
cycler==0.12.1
kiwisolver==1.4.5
python-dateutil==2.9.0.post0
pytz==2024.1
packaging==24.1

# Jupyter stack (compatible with Python 3.8)
ipython==7.34.0
ipykernel==6.29.4
jupyterlab==3.6.7
jupyterlab-server==2.25.1
jupyter-client==8.6.0
jupyter-core==5.7.2
notebook==6.4.13
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
jupyterlab-pygments==0.3.0
traitlets==5.14.2
async-generator==1.10
nest-asyncio==1.6.0

# HTML / XML / rendering
bleach==6.1.0
defusedxml==0.7.1
MarkupSafe==2.1.5
Jinja2==3.1.4
mistune==2.0.5

# Network / utils
requests==2.32.3
certifi==2024.7.4
urllib3==2.2.3
chardet==5.2.0
idna==3.7

# Console / CLI tools
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0

# Security
argon2-cffi==23.1.0
cffi==1.16.0
pycparser==2.22
PyJWT==2.8.0

# Visualization + utils
Pillow==10.4.0
Pygments==2.18.0
pyparsing==3.1.2

# ZMQ (ARM64 wheels)
pyzmq==25.1.1
tornado==6.4.1

# Jupyter helper libs
entrypoints==0.4
prometheus-client==0.21.0
Send2Trash==1.8.2
terminado==0.18.1
testpath==0.6.0
wcwidth==0.2.13
webencodings==0.5.1

# Misc
decorator==5.1.1
inflect==7.0.0
jsonschema==4.22.0
pandocfilters==1.5.0

# WSGI server
uWSGI==2.0.21
